changed folder structure and added APIRouter routes
