+++ dependencies.py
@@ -1,0 +1,7 @@
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
