+++ requirement.txt
@@ -1,51 +1,6 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: linux-64
-_libgcc_mutex=0.1=conda_forge
-_openmp_mutex=4.5=1_gnu
-ca-certificates=2020.12.5=ha878542_0
-certifi=2020.12.5=py39hf3d152e_1
-click=7.1.2=pyh9f0ad1d_0
-fastapi=0.63.0=pyhd8ed1ab_0
-greenlet=1.0.0=py39he80948d_0
-h11=0.12.0=pyhd8ed1ab_0
-httptools=0.1.1=py39h3811e60_2
-jedi=0.18.0=py39hf3d152e_2
-krb5=1.17.2=h926e7f8_0
-ld_impl_linux-64=2.35.1=hea4e1c9_2
-libedit=3.1.20191231=he28a2e2_2
-libffi=3.3=h58526e2_2
-libgcc-ng=9.3.0=h2828fa1_18
-libgomp=9.3.0=h2828fa1_18
-libpq=13.1=hfd2b0eb_2
-libstdcxx-ng=9.3.0=h6de172a_18
-libuv=1.41.0=h7f98852_0
-msgpack-python=1.0.2=py39h1a9c180_1
-ncurses=6.2=h58526e2_4
-openssl=1.1.1j=h7f98852_0
-parso=0.8.1=pyhd8ed1ab_0
-pip=21.0.1=pyhd8ed1ab_0
-psycopg2=2.8.6=py39h3811e60_2
-pydantic=1.8.1=py39h3811e60_1
-pynvim=0.4.3=py39hf3d152e_0
-python=3.9.2=hffdb5ce_0_cpython
-python-dotenv=0.15.0=pyhd8ed1ab_0
-python_abi=3.9=1_cp39
-pyyaml=5.4.1=py39h3811e60_0
-readline=8.0=he28a2e2_2
-setuptools=49.6.0=py39hf3d152e_3
-sqlalchemy=1.4.2=py39h3811e60_0
-sqlite=3.34.0=h74cdb3f_0
-starlette-base=0.13.6=py_0
-tk=8.6.10=h21135ba_1
-typing-extensions=3.7.4.3=0
-typing_extensions=3.7.4.3=py_0
-tzdata=2021a=he74cb21_0
-uvicorn=0.13.4=py39hf3d152e_0
-uvloop=0.15.2=py39h3811e60_0
-watchgod=0.6=py_0
-websockets=8.1=py39h3811e60_3
-wheel=0.36.2=pyhd3deb0d_0
-xz=5.2.5=h516909a_1
-yaml=0.2.5=h516909a_0
-zlib=1.2.11=h516909a_1010
+fastapi==0.63.0
+greenlet==1.0.0
+pydantic==1.8.1
+SQLAlchemy==1.4.6
+starlette==0.13.6
+typing-extensions==3.7.4.3
