+++ routers/auction.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get('/')
+async def get_auction():
+    return 'auction'
